[Single/Double.Matrix] Remove redundant calculation of abs in frobenius norm
The formulas found in various references do not have abs. Furthermore,
when multiplying A*A^T or A^T*A (for the frobenius norm we only care
about the diagonal so it doesn't matter which product is used) the
i-ith diagonal element of the final matrix comes from the multiplication
of line i with itself. Therefore, all numbers involved will be multiplied
with themselves resulting in numbers that are always non-negative.

Signed-off-by: Alexander Karatarakis <alex@karatarakis.com>