Native: automatically use native provider if available #314, new Control.TryXY alternatives
Control.TryUseNativeMKL() etc as an alternative way to use a specific
provider if it is available, but you're not interested in why if not.

Control.TryUseNative() tries to use any known native provider in an
undefined (but reasonable) order.

Changed the default behavior to try to load any known native provider
available automatically instead of falling back to managed straight away.

Added basic inline-xml docs